Initial utils.c platform code refactor
