+++ src/core/utils.c
@@ -1,7 +1,3 @@
-#if __APPLE__ && __MACH__
-#include <fcntl.h>	/* For fcntl().	*/
-#endif	/* Apple Macintosh */
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
@@ -197,72 +193,3 @@ void print_header(FILE *f, const char *name)
 	"This is free software; see the source for copying conditions.\n"
 	"\n", name);
 }
-
-#if __APPLE__ && __MACH__
-
-/* This function is a _rough_ approximation of fdatasync(2). */
-int fdatasync(int fd)
-{
-	return fcntl(fd, F_FULLFSYNC);
-}
-
-/* This function is a _rough_ approximation of posix_fadvise(2). */
-int posix_fadvise(int fd, off_t offset, off_t len, int advice)
-{
-	UNUSED(offset);
-	UNUSED(len);
-	switch (advice) {
-	case POSIX_FADV_SEQUENTIAL:
-		return fcntl(fd, F_RDAHEAD, 1);
-	case POSIX_FADV_DONTNEED:
-		return fcntl(fd, F_NOCACHE, 1);
-	default:
-		assert(0);
-	}
-}
-
-#endif	/* Apple Macintosh */
-
-#if (__APPLE__ && __MACH__) || defined(__OpenBSD__)
-
-void msleep(double wait_ms)
-{
-	assert(!usleep(wait_ms * 1000));
-}
-
-#else	/* Apple Macintosh / OpenBSD */
-
-#include <time.h> /* For clock_gettime() and clock_nanosleep(). */
-
-void msleep(double wait_ms)
-{
-	struct timespec req;
-	int ret;
-
-	assert(!clock_gettime(CLOCK_MONOTONIC, &req));
-
-	/* Add @wait_ms to @req. */
-	if (wait_ms > 1000) {
-		time_t sec = wait_ms / 1000;
-		wait_ms -= sec * 1000;
-		assert(wait_ms > 0);
-		req.tv_sec += sec;
-	}
-	req.tv_nsec += wait_ms * 1000000;
-
-	/* Round @req up. */
-	if (req.tv_nsec >= 1000000000) {
-		ldiv_t result = ldiv(req.tv_nsec, 1000000000);
-		req.tv_sec += result.quot;
-		req.tv_nsec = result.rem;
-	}
-
-	do {
-		ret = clock_nanosleep(CLOCK_MONOTONIC, TIMER_ABSTIME,
-			&req, NULL);
-	} while (ret == EINTR);
-
-	assert(ret == 0);
-}
-
-#endif	/* Apple Macintosh / OpenBSD */
